+++ functions/package.json
@@ -13,6 +13,7 @@
   },
   "main": "index.js",
   "dependencies": {
+    "axios": "^1.2.1",
     "firebase-admin": "^11.3.0",
     "firebase-functions": "^4.1.1"
   },
